requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.35.0
botocore==1.35.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cohere==5.8.1
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
experta==1.9.4
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==1.2
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jax==0.4.31
jaxlib==0.4.31
jmespath==1.0.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
msgpack==1.0.8
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
parameterized==0.9.0
pillow==10.4.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
schema==0.6.7
scipy==1.14.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
sqlparse==0.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tokenizers==0.20.0
tqdm==4.66.5
Twisted==24.3.0
txaio==23.1.1
types-PyYAML==6.0.12.20240724
types-requests==2.32.0.20240712
typing_extensions==4.12.0
urllib3==2.2.2
uuid==1.30
Werkzeug==3.0.3
whitenoise==6.6.0
wrapt==1.16.0
zope.interface==6.4.post2
